MUMPS get rid of uneffective omp directives
The problem is the same as static variables for which the scope is the translation unit (e.g. the shared library in which they are compiled).
So the only way to set the 'omp_num_threads' LOCAL variable (et simila) for a specific translation unit (in this case the mumps and/or the BLAS library) is to have a function FROM THAT translation unit that does the job.
This function can access the omp local variable because they both share the same translation unit.
Otherwise, you can superimpose the number of threads using the envvar OMP_NUM_THREADS. Then the translations units...
- that DO NOT specify the thread number will adopt OMP_NUM_THREADS
- that DO specify the thread number won't be affected